requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0,<4.0
python-dotenv>=0.19.0
gunicorn>=20.1.0